pyproject.toml
[project]
name = "prepdir"
version = "0.10.0"
description = "Directory traversal utility to prepare project contents for review"
readme = "README.md"
authors = [
    {name = "eyecantell", email = "paul@pneuma.solutions"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Topic :: Text Processing :: Markup",
    "Development Status :: 4 - Beta",
]
keywords = [
    "ai", 
    "artificial intelligence", 
    "code review", 
    "directory traversal", 
    "file content", 
    "project documentation", 
    "code sharing", 
    "developer tools", 
    "large language models",
    "llm",
    "project structure"
]
requires-python = ">=3.8"
dependencies = [
    "typing-extensions>=4.7.1; python_version < '3.11'",
    "importlib-resources>=5.0; python_version < '3.9'",
    "pyyaml>=6.0",
    "dynaconf>=3.2.6",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "coverage>=7.2.7",
]

[project.scripts]
prepdir = "prepdir.main:main"

[tool.ruff]
line-length = 120

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
package-dir = "src"
includes = ["src/prepdir", "src/prepdir/config.yaml"]

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Repository = "https://github.com/eyecantell/prepdir"
Issues = "https://github.com/eyecantell/prepdir/issues"
Documentation = "https://github.com/eyecantell/prepdir#readme"